+++ website-frontend/components.json
@@ -7,7 +7,7 @@
 		"baseColor": "slate"
 	},
 	"aliases": {
-		"components": "$lib/@shadcn-svelte/components",
+		"components": "$lib/@shadcn-svelte/",
 		"utils": "$lib/@shadcn-svelte/utils"
 	},
 	"typescript": true
