+++ website-frontend/src/lib/@shadcn-svelte/ui/carousel/carousel-content.svelte
