+++ website-frontend/src/lib/@shadcn-svelte/ui/carousel/carousel-item.svelte
