+++ website-frontend/src/lib/@shadcn-svelte/ui/carousel/carousel-previous.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import ArrowRight from "lucide-svelte/icons/arrow-right";
+	import ArrowLeft from "lucide-svelte/icons/arrow-left";
 	import type { VariantProps } from "tailwind-variants";
 	import { getEmblaContext } from "./context.js";
 	import { cn } from "$lib/@shadcn-svelte/utils.js";
@@ -7,7 +7,7 @@
 		Button,
 		type Props,
 		type buttonVariants,
-	} from "$lib/@shadcn-svelte/components/ui/button/index.js";
+	} from "$lib/@shadcn-svelte//ui/button/index.js";
 
 	type $$Props = Props;
 
@@ -15,8 +15,9 @@
 	export { className as class };
 	export let variant: VariantProps<typeof buttonVariants>["variant"] = "outline";
 	export let size: VariantProps<typeof buttonVariants>["size"] = "icon";
-	const { orientation, canScrollNext, scrollNext, handleKeyDown } =
-		getEmblaContext("<Carousel.Next/>");
+
+	const { orientation, canScrollPrev, scrollPrev, handleKeyDown } =
+		getEmblaContext("<Carousel.Previous/>");
 </script>
 
 <Button
@@ -25,15 +26,15 @@
 	class={cn(
 		"absolute h-8 w-8 touch-manipulation rounded-full",
 		$orientation === "horizontal"
-			? "-right-12 top-1/2 -translate-y-1/2"
-			: "-bottom-12 left-1/2 -translate-x-1/2 rotate-90",
+			? "-left-12 top-1/2 -translate-y-1/2"
+			: "-top-12 left-1/2 -translate-x-1/2 rotate-90",
 		className
 	)}
-	disabled={!$canScrollNext}
-	on:click={scrollNext}
+	disabled={!$canScrollPrev}
+	on:click={scrollPrev}
 	on:keydown={handleKeyDown}
 	{...$$restProps}
 >
-	<ArrowRight class="h-4 w-4" />
-	<span class="sr-only">Next slide</span>
+	<ArrowLeft class="h-4 w-4" />
+	<span class="sr-only">Previous slide</span>
 </Button>
