+++ website-frontend/src/lib/@shadcn-svelte/ui/tabs/index.ts
