+++ website-frontend/src/lib/@shadcn-svelte/ui/tabs/tabs-trigger.svelte
